fix: avoid importing self as a related model
This patch resolves an issue with recursive model types. The SDK
generator used to blindly assume that all refs in the schema are
external models, i.e. it didn't cover the case where a model recursively
references itself. This caused an issue with the generation of
`SearchResultDTO`. While the preferred solution in the past appears to
have been creating duplicate types to avoid recursive references, it's a
good idea for the generator to not fail in such cases as well.